refactor: Move InvalidCurrentValueError to ~/errors
