+++ packages/server/src/internals/data-source-helper/index.ts
@@ -1,4 +1,3 @@
-export { InvalidCurrentValueError } from './common'
 export { createMutation } from './mutation'
 export { createRequest } from './request'
 export {
